move /post and /events to api directory
